+++ src/lib.rs
@@ -22,4 +22,4 @@ pub mod error;
 #[cfg(feature = "serde_impl")]
 pub use error::ProblemDetail;
 
-pub use error::{ErrCode, ErrorCode};
+pub use error::{ErrCode, ErrorCode, HttpError};